refactor: orderBy를 별개의 클래스로 분리
