+++ src/main/java/com/sprint/sb06deokhugamteam01/dto/book/request/BookCreateRequest.java
@@ -1,7 +1,7 @@
 package com.sprint.sb06deokhugamteam01.dto.book.request;
 
 import lombok.Builder;
-import com.sprint.sb06deokhugamteam01.domain.Book;
+import com.sprint.sb06deokhugamteam01.domain.book.Book;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
