+++ src/main/java/com/sprint/sb06deokhugamteam01/repository/BookRepository.java
@@ -1,7 +1,6 @@
 package com.sprint.sb06deokhugamteam01.repository;
 
-import com.sprint.sb06deokhugamteam01.domain.Book;
-import org.springframework.data.jpa.repository.JpaRepository;
+import com.sprint.sb06deokhugamteam01.domain.book.Book;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
